Rename checker builder file
