+++ portal/server/src/list_checker_builder.ts
